+++ client/src/components/Nav/ExportConversation/index.ts
@@ -1,2 +1,1 @@
-export { default as ExportConversation } from './ExportConversation';
 export { default as ExportModal } from './ExportModal';
