+++ client/src/hooks/Conversations/index.ts
@@ -8,3 +8,4 @@ export { default as useDebouncedInput } from './useDebouncedInput';
 export { default as useNavigateToConvo } from './useNavigateToConvo';
 export { default as useSetIndexOptions } from './useSetIndexOptions';
 export { default as useParameterEffects } from './useParameterEffects';
+export { default as useExportConversation } from './useExportConversation';
